+++ src/main/java/com/work/crm/CRM/domain/person/controler/PersonControler.java
@@ -39,7 +39,7 @@ public class PersonControler {
     }
 
     @PutMapping(path = "/persons/{id}")
-    ResponseEntity<?> updateCountry(@PathVariable int id, @RequestBody Person toUpdate)
+    ResponseEntity<?> updateCountry(@PathVariable long id, @RequestBody Person toUpdate)
     {
         if (personRepository.existsById(id)) {
             personRepository.save(toUpdate);
