+++ src/main/java/com/work/crm/CRM/domain/rltAddressPerson/entity/RLTAddressPersonRepository.java
@@ -10,7 +10,7 @@ public interface RLTAddressPersonRepository {
     List<RLTAddressPerson> findAll();
     Page<RLTAddressPerson> findAll(Pageable page);
     Optional<RLTAddressPerson> findById(Integer i);
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
     RLTAddressPerson save(RLTAddressPerson entity);
 //    List<RLTAddressPerson> saveAll(List<RLTAddressPerson> entities);
